requirements.txt
Django==4.2
djangorestframework==3.15.2
Pillow==9.0.0
